==> notebooks/01_preprocessing.ipynb <==
# Preprocessing

## Import libraries

import pandas as pd
import numpy as np

## Get the datasets

train = pd.read_json('../datasets/train.json')
test = pd.read_json('../datasets/test.json')

## Preprocessing the `interes_level`

train['interest_level'].unique()

train['interest_level'] = train['interest_level'].map({'low' : 0, 'medium' : 1, 'high' : 2})

train['interest_level'].unique()

## Cleaning of emissions

low_perc, high_perc = np.percentile(train['price'], [1, 99])
low_perc, high_perc

train = train[(train['price'] > low_perc) & (train['price'] < high_perc)]
test = test[(test['price'] > low_perc) & (test['price'] < high_perc)]

idx_bathrooms = test['bathrooms'].nlargest(5).index
test = test.drop(idx_bathrooms)

test.describe()[['bathrooms', 'bedrooms', 'price']]

## Save dataframe

train.to_pickle('../datasets/train_01.pkl')
test.to_pickle('../datasets/test_01.pkl')
