==> notebooks/02_features.ipynb <==
# Feature engineering

## Import libraries

import pandas as pd
from collections import Counter

## Get the datasets

train = pd.read_pickle('../datasets/train_01.pkl')
test = pd.read_pickle('../datasets/test_01.pkl')

## Engineering

### Clean the column `features`

train['features'] = train['features'].apply(lambda x : [item.replace('[', '').replace(']', '').replace("'", "").replace('"', '').replace(' ', '') for item in x])

train['features']

### Let's count the top 20 signs

all_features = []

for item in train['features']:
    all_features.extend(item)

print("Number of unique features: ", len(set(all_features)))

cnt = Counter(all_features)
cnt.most_common(20)

### Creating a one-hot tables

top_features = [i[0] for i in cnt.most_common(20)]

for feature in top_features:
    train[feature] = train['features'].apply(lambda x : 1 if feature in x else 0)
    test[feature] = test['features'].apply(lambda x : 1 if feature in x else 0)

feature_list = ['bathrooms', 'bedrooms'] + top_features

oneHot_train = train[feature_list + ['price']]
oneHot_test = test[feature_list + ['price']]

## Saving the resulting tables

oneHot_train.to_pickle('../datasets/train_02_onehot.pkl')
oneHot_test.to_pickle('../datasets/test_02_onehot.pkl')
