==> notebooks/04_regularization.ipynb <==
# Implementing a regularization models

## Import libraries

import pandas as pd
from sklearn.linear_model import Ridge, Lasso, ElasticNet

import sys
sys.path.append('../')

from src.models import MyRidge, MyLasso, MyElasticNet
from src.utils import ModelComparator

## Get the dataframes

train = pd.read_pickle('../datasets/train_02_onehot.pkl')
test = pd.read_pickle('../datasets/test_02_onehot.pkl')

X_train = train.drop(['price'], axis=1)
y_train = train['price']
X_test = test.drop(['price'], axis=1)
y_test = test['price']

## Test a regularization models

comparator = ModelComparator(X_train, y_train, X_test, y_test)

comparator.compare_with_library(MyRidge(), Ridge())

comparator.compare_with_library(MyLasso(), Lasso())

comparator.compare_with_library(MyElasticNet(), ElasticNet())

## Save a results

comparator.evaluate_model(MyLasso(), "Lasso_default")

comparator.evaluate_model(MyRidge(), "Ridge_default")

comparator.evaluate_model(MyElasticNet(), "ElasticNet_default")

comparator.save_results()
