==> notebooks/05_normalization.ipynb <==
## Normalization

## Import a libraies

import pandas as pd
import numpy as np
from sklearn.linear_model import LinearRegression, Ridge, Lasso, ElasticNet
from sklearn.preprocessing import MinMaxScaler, StandardScaler

import sys
sys.path.append('../')

from src.models import MyRidge, MyLasso, MyElasticNet, LinearRegressionSGD
from src.utils import MyMinMaxScaler, MyStandardScaler
from src.utils import ModelComparator

## Get a dataframe

train = pd.read_pickle('../datasets/train_02_onehot.pkl')
test = pd.read_pickle('../datasets/test_02_onehot.pkl')

X_train = train.drop(['price'], axis=1)
y_train = train['price']
X_test = test.drop(['price'], axis=1)
y_test = test['price']

## Comparison of normalization implementations

### StandartScaler

my_scaler = MyStandardScaler()
lib_scaler = StandardScaler()

(my_scaler.fit_transform(X_train) - lib_scaler.fit_transform(X_train)).sum()

np.allclose(my_scaler.fit_transform(X_train), lib_scaler.fit_transform(X_train))

### MinMaxScaler

my_scaler = MyMinMaxScaler()
lib_scaler = MinMaxScaler()

(my_scaler.fit_transform(X_train) - lib_scaler.fit_transform(X_train)).sum()

np.allclose(my_scaler.fit_transform(X_train), lib_scaler.fit_transform(X_train))

## Training of linear models using a scaling

### Using MinMaxScaler

minmax_comparator = ModelComparator(X_train, y_train, X_test, y_test, MyMinMaxScaler())

minmax_comparator.evaluate_model(LinearRegressionSGD(), "LinearSGD_minmax")

minmax_comparator.evaluate_model(MyLasso(), "Lasso_minmax")

minmax_comparator.evaluate_model(MyRidge(), "Ridge_minmax")

minmax_comparator.evaluate_model(MyElasticNet(), "ElasticNet_minmax")

minmax_comparator.save_results()

### Using StandartScaler

standart_comparator = ModelComparator(X_train, y_train, X_test, y_test, MyStandardScaler())

standart_comparator.evaluate_model(LinearRegressionSGD(), "LinearSGD_standart")

standart_comparator.evaluate_model(MyLasso(), "Lasso_standart")

standart_comparator.evaluate_model(MyRidge(), "Ridge_standart")

standart_comparator.evaluate_model(MyElasticNet(), "ElasticNet_standart")

standart_comparator.save_results()
