==> notebooks/06_poly_overfit.ipynb <==
# Checking the overfit of the model

## Import libraries

import pandas as pd
from sklearn.preprocessing import PolynomialFeatures
from sklearn.linear_model import LinearRegression, Lasso, Ridge, ElasticNet
import sys
sys.path.append('../')

from src.models import MyRidge, MyLasso, MyElasticNet, LinearRegressionSGD
from src.utils import ModelComparator

## Get the dataframes

train = pd.read_pickle('../datasets/train_02_onehot.pkl')
test = pd.read_pickle('../datasets/test_02_onehot.pkl')

X_train = train[['bathrooms', 'bedrooms']]
y_train = train['price']
X_test = test[['bathrooms', 'bedrooms']]
y_test = test['price']

## Train models with polynomical features

poly = PolynomialFeatures(degree=10)

X_poly_train = poly.fit_transform(X_train)
X_poly_test = poly.transform(X_test)

comparator = ModelComparator(X_train, y_train, X_test, y_test, poly)

comparator.evaluate_model(LinearRegression(), "LinearSGD_poly")

comparator.evaluate_model(Lasso(), "Lasso_poly")

comparator.evaluate_model(Ridge(), "Ridge_poly")

comparator.evaluate_model(ElasticNet(), "ElasticNet_poly")

comparator.get_results()[1]

comparator.save_results()
