==> notebooks/07_naive.ipynb <==
# Naive models

## Import Libraries

import pandas as pd
import numpy as np
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
import sys
sys.path.append('../')
from src.utils import ModelComparator

## Get the dataframe

train = pd.read_pickle('../datasets/train_02_onehot.pkl')
test = pd.read_pickle('../datasets/test_02_onehot.pkl')

y_train = train['price']
y_test = test['price']

## Implementation of naive models

median_train = y_train.median()
mean_train = y_train.mean()

median_test = y_test.median()
mean_test = y_test.mean()

(median_test, mean_test), (median_train, mean_train)

y_train_mean = np.full((len(y_train), 1), mean_train)
y_train_median = np.full((len(y_train), 1), median_train)

y_test_mean = np.full((len(y_test), 1), mean_test)
y_test_median = np.full((len(y_test), 1), median_test)

median_train_mae = mean_absolute_error(y_train, y_train_median)
median_train_rmse = np.sqrt(mean_squared_error(y_train, y_train_median))
median_train_r2 = r2_score(y_train, y_train_median)

mean_train_mae = mean_absolute_error(y_train, y_train_mean)
mean_train_rmse = np.sqrt(mean_squared_error(y_train, y_train_mean))
mean_train_r2 = r2_score(y_train, y_train_mean)

median_test_mae = mean_absolute_error(y_test, y_test_median)
median_test_rmse = np.sqrt(mean_squared_error(y_test, y_test_median))
median_test_r2 = r2_score(y_test, y_test_median)

mean_test_mae = mean_absolute_error(y_test, y_test_mean)
mean_test_rmse = np.sqrt(mean_squared_error(y_test, y_test_mean))
mean_test_r2 = r2_score(y_test, y_test_mean)

comparator = ModelComparator(None, None, None, None, None)

comparator._update_results(
    comparator.results_mae,
    "Naive_mean",
    mean_train_mae,
    mean_test_mae
)

comparator._update_results(
    comparator.results_rmse,
    "Naive_mean",
    mean_train_rmse,
    mean_test_rmse
)

comparator._update_results(
    comparator.results_r2,
    "Naive_mean",
    mean_train_r2,
    mean_test_r2
)

comparator._update_results(
    comparator.results_mae,
    "Naive_median",
    median_train_mae,
    median_test_mae
)

comparator._update_results(
    comparator.results_rmse,
    "Naive_median",
    median_train_rmse,
    median_test_rmse
)

comparator._update_results(
    comparator.results_r2,
    "Naive_median",
    median_train_r2,
    median_test_r2
)

comparator.save_results()
